fix(M12): listByBbox zod accepts float zoom (MapLibre returns fractional zoom)
Ming's console diagnostic isolated the root cause: every call to
toilet.listByBbox came back with isFetching=false, isSuccess=false,
data=undefined, and isError=true — a BAD_REQUEST from the server's
zod guard rejecting `zoom` because MapLibre's map.getZoom() returns
a float mid-gesture (e.g. 7.086607388906396) while the schema
required .int().

Rejection chain that masked the bug:
  1. Client sent e.g. { zoom: 7.0866… } on every moveend
  2. Server zod parse failed → TRPCError BAD_REQUEST
  3. React Query recorded error + set isError=true
  4. In prod (what Ming tested against), the tRPC loggerLink is
     gated on NODE_ENV==='development' so the error surfaced
     nowhere — no console.error, no DevTools indication, clean
     network 200 responses with an error payload buried inside
  5. Users see an empty map outside Tokyo (which had cached data)

Fix: drop .int() from the zoom field. The server doesn't consume
zoom for anything except future logging, so integer-ness was a
cosmetic precondition; the min(0).max(22) range check still holds.

Verified: pnpm typecheck / lint / test (132/132) all green.
Post-deploy Ming should hard-reload toirepo.com and see markers
populate in Beijing / NYC / London viewports.

+++ src/server/api/routers/toilet/listByBbox.ts
@@ -39,7 +39,14 @@ export const ListByBboxInputSchema = z.object({
   // queries below zoom 3 to avoid whole-earth envelopes that can't
   // fit inside the limit. Keeping it in the input makes
   // server-side logging/metrics aware of view context if needed.
-  zoom: z.number().int().min(0).max(22),
+  //
+  // NOT .int(): MapLibre's map.getZoom() returns a fractional float
+  // during any transition (e.g. 7.086607388906396 mid-pinch). An
+  // .int() guard rejected every single client query with BAD_REQUEST
+  // — React Query saw isSuccess=false, data undefined, no markers
+  // rendered anywhere, and in prod builds the tRPC loggerLink is
+  // disabled so the error was entirely silent.
+  zoom: z.number().min(0).max(22),
   // 5000 is the ceiling the original M11 schema tolerated when
   // MapLibre's cluster aggregation was benchmarked against Tokyo's
   // 10k rows. A typical mid-zoom city viewport returns well under
